moved the theme to its own custom provider
